Kerberos: agents without certs, with SPNEGO
This *prototype* change demonstrates a minimal set of edits to support
SPNEGO/GSSAPI/Kerberos authentication of agents to masters instead of agent
certificates. It is not meant to be a shippable patch. There are two primary
changes:

1) Add SPNEGO/gssapi support to Puppet::Network::HTTP::Connection.

2) Add an agent option to run without a certificate. For now stub out the
   certificate creation and checking logic.

In addition to all the TODOs in the code this change is also missing tests.